New unique: Nawab Begum
A very early-game hobgoblin unique (currently at D2-4); about
twice as tough as a hobgoblin, but placing with a mob of goblin
& hobgoblin minions, and with the ability to buff them (with
Battlecry) and hurl them at the player.

As a reward for killing her, she has a random piece of jewellery
(since she's a wealthy nabob!) - this is an experiment. If she berserks
too often, it can be removed or replaced with gold, something like that.

(The code for goblin tossing is extremely bad and shameful.
I am sorry. It is a first draft.)

She intentionally has no montitle, since, like Grinder or the Enchantress,
"Nawab Begum" is actually a title itself.