Remove porridge
Like blood for non-vp, this was no longer needed now that Gozag's
potion petition doesn't offer it. Again, it's redundant with other
food types, and pretty unexciting to boot. I mean... porridge?

I've replaced it with stacks of royal jelly in most cases, which
is a significant boost for carnivores (who got nothing at all from
porridge). If this is considered undesirable, they can be swapped
out for bread rations instead.